requirements.txt
django-celery-beat>=2.6.0
django-celery-results~=2.5.1